+++ src/main/java/de/maxhenkel/shulkerbox/Events.java
@@ -1,126 +1,92 @@
 package de.maxhenkel.shulkerbox;
 
 import de.maxhenkel.shulkerbox.gui.ShulkerboxContainer;
-import net.minecraft.block.Block;
-import net.minecraft.block.ShulkerBoxBlock;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.inventory.container.INamedContainerProvider;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Hand;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.util.SoundCategory;
+import net.minecraft.util.SoundEvents;
 import net.minecraftforge.common.util.FakePlayer;
+import net.minecraftforge.event.entity.player.PlayerContainerEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
-import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.network.NetworkHooks;
-
-import javax.annotation.Nullable;
 
 @Mod.EventBusSubscriber(modid = Main.MODID)
 public class Events {
 
     @SubscribeEvent
-    public void onRightClick(BlockEvent.EntityPlaceEvent event) {
-        if (event.isCanceled()) {
-            return;
-        }
-
-        if (!(event.getEntity() instanceof PlayerEntity)) {
-            return;
-        }
-
-        PlayerEntity player = (PlayerEntity) event.getEntity();
-
-        if (player instanceof FakePlayer) {
-            return;
-        }
-
-        if (!(event.getPlacedBlock().getBlock() instanceof ShulkerBoxBlock)) {
-            return;
-        }
-
-        ItemStack stack = getShulkerBox(player);
-        if (stack == null) {
-            return;
-        }
-
-        if (Config.onlySneakPlace) {
-            if (!player.func_225608_bj_()) {
-                displayGUI(player, stack);
-                event.setCanceled(true);
-            }
-        }
-
-    }
-
-    @SubscribeEvent
-    public void onRightClick(PlayerInteractEvent.RightClickItem event) {
+    public void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
         if (event.isCanceled()) {
             return;
         }
 
         PlayerEntity player = event.getPlayer();
+        Hand hand = event.getHand();
 
         if (player instanceof FakePlayer) {
             return;
         }
 
-        if (!isShulkerBox(event.getItemStack())) {
+        if (hand.equals(Hand.OFF_HAND) && Utils.getShulkerBox(player, Hand.MAIN_HAND) != null) {
+            event.setCanceled(true);
             return;
         }
 
-        ItemStack stack = getShulkerBox(player);
-        if (stack != null) {
-            displayGUI(player, stack);
-        }
-    }
-
-    public static ItemStack getShulkerBox(PlayerEntity player) {
-        ItemStack stack = player.getHeldItem(Hand.MAIN_HAND);
-        if (isShulkerBox(stack)) {
-            return stack;
-        }
-        stack = player.getHeldItem(Hand.OFF_HAND);
-        if (isShulkerBox(stack)) {
-            return stack;
-        }
-        return null;
-    }
-
-    private static boolean isShulkerBox(ItemStack stack) {
+        ItemStack stack = Utils.getShulkerBox(player, hand);
         if (stack == null) {
-            return false;
+            return;
         }
 
-        Item item = stack.getItem();
-
-        if (Block.getBlockFromItem(item) instanceof ShulkerBoxBlock) {
-            return true;
+        if (Config.SERVER.onlySneakPlace.get()) {
+            if (!player.func_225608_bj_()) {
+                Utils.openShulkerBox(player, stack);
+                event.setCanceled(true);
+            }
         }
-
-        return false;
     }
 
-    private void displayGUI(PlayerEntity player, ItemStack stack) {
-        if (!player.world.isRemote && player instanceof ServerPlayerEntity) {
-            NetworkHooks.openGui((ServerPlayerEntity) player, new INamedContainerProvider() {
-                @Nullable
-                @Override
-                public Container createMenu(int id, PlayerInventory playerInventory, PlayerEntity playerEntity) {
-                    return new ShulkerboxContainer(id, playerInventory, new ShulkerBoxInventory(player, stack));
-                }
+    @SubscribeEvent
+    public void onRightClickItem(PlayerInteractEvent.RightClickItem event) {
+        if (event.isCanceled()) {
+            return;
+        }
 
-                @Override
-                public ITextComponent getDisplayName() {
-                    return new TranslationTextComponent(stack.getTranslationKey());
-                }
-            });
+        PlayerEntity player = event.getPlayer();
+        Hand hand = event.getHand();
+
+        if (player instanceof FakePlayer) {
+            return;
+        }
+
+        if (hand.equals(Hand.OFF_HAND) && Utils.getShulkerBox(player, Hand.MAIN_HAND) != null) {
+            event.setCanceled(true);
+            return;
+        }
+
+        if (player.openContainer instanceof ShulkerboxContainer) {
+            return;
+        }
+
+        ItemStack stack = Utils.getShulkerBox(player, hand);
+        if (stack != null) {
+            Utils.openShulkerBox(player, stack);
+        }
+    }
+
+    @SubscribeEvent
+    public void openShulkerBox(PlayerContainerEvent.Open event) {
+        if (event.getContainer() instanceof ShulkerboxContainer) {
+            PlayerEntity player = event.getPlayer();
+            player.world.playSound(null, player.func_226277_ct_(), player.func_226278_cu_(), player.func_226281_cx_(), SoundEvents.BLOCK_SHULKER_BOX_OPEN, SoundCategory.BLOCKS, 0.5F, player.world.rand.nextFloat() * 0.1F + 0.9F);
+        }
+    }
+
+    @SubscribeEvent
+    public void closeShulkerBox(PlayerContainerEvent.Close event) {
+        if (event.getContainer() instanceof ShulkerboxContainer) {
+            PlayerEntity player = event.getPlayer();
+            player.world.playSound(null, player.func_226277_ct_(), player.func_226278_cu_(), player.func_226281_cx_(), SoundEvents.BLOCK_SHULKER_BOX_CLOSE, SoundCategory.BLOCKS, 0.5F, player.world.rand.nextFloat() * 0.1F + 0.9F);
         }
     }